+++ src/services/dynamodb.js
@@ -1,0 +1,4 @@
+import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
+import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
+
+export const ddbDoc = DynamoDBDocumentClient.from(new DynamoDBClient({}));
