+++ src/services/openAI.js
@@ -1,0 +1,6 @@
+import 'dotenv/config';
+import OpenAI from 'openai';
+
+export const openai = new OpenAI({
+  apiKey: process.env.OPEN_AI_API_KEY,
+});
